Title: 최솟값 3개, Time: 1611ms, Memory: 42MB, Status: Passed - CodeTree

+++ 241024/최솟값 3개/top-3-smallest-number.java
@@ -6,15 +6,15 @@ public class Main {
         Scanner sc = new Scanner(System.in);
 
         int n = sc.nextInt();
-        PriorityQueue<Integer> pq = new PriorityQueue<>();
+        PriorityQueue<Long> pq = new PriorityQueue<>();
 
         for (int i=0; i<n; i++) {
-            pq.add(sc.nextInt());
+            pq.add((long)sc.nextInt());
 
             if (pq.size() >= 3) {
-                int num1 = pq.poll();
-                int num2 = pq.poll();
-                int num3 = pq.poll();
+                Long num1 = pq.poll();
+                Long num2 = pq.poll();
+                Long num3 = pq.poll();
                 System.out.println(num1 * num2 * num3);
 
                 pq.add(num1);
